feat: Added component MessageSearchBlock
